refactor: liquidation engine architecture migration to Rust core
Addresses all maintainer feedback from PR #4077 detailed review:

- Migrated liquidation logic from Cython to Rust-only implementation
- Removed SimulatedExchange._process_margin_liquidations() and OrderMatchingEngine.liquidate_open_positions() from engine.pyx/engine.pxd
- Tests now drive Rust path via BacktestEngine pyo3 bindings

- Removed liquidated_accounts dedupe set; added retrigger regression test
- OrderSubmitted event now emitted in full Initialized->Submitted->Accepted->Filled sequence
- Account ID panic path fixed via explicit tuple carry and account_ids mapping
- Uses OrderCore::closing_side(pos.side) instead of manual match
- No OrderSide::NoOrderSide fallback in liquidation path

- Removed Unicode characters (arrows, dashes) from test/example files
- Test uses free functions with fixtures (no TestLiquidationEngine class)
- liquidation_trigger_ratio aligned to f64 throughout (Rust/Python/config)
- BacktestEngine.add_venue() docstring documents all three liquidation params
- Example moved to examples/backtest/ and uses add_venue liquidation flags

- TimeInForce::Ioc for synthetic close orders (immediate execution)
- positions_open hoisted and pre-bucketed by settlement_currency
- Cross-margin assumption documented in process_liquidations docstring
- All 55 pre-commit hooks pass (cargo fmt, clippy, mypy, ruff, etc.)

+++ crates/backtest/src/config.rs
@@ -434,8 +434,8 @@ pub struct BacktestVenueConfig {
     liquidation_enabled: bool,
     /// The ratio of equity to maintenance margin at which liquidation is triggered.
     /// A value of 1.0 means liquidation triggers when equity <= maintenance_margin.
-    #[builder(default = Decimal::ONE)]
-    liquidation_trigger_ratio: Decimal,
+    #[builder(default = 1.0)]
+    liquidation_trigger_ratio: f64,
     /// If open orders should be canceled before closing positions during liquidation.
     #[builder(default = true)]
     liquidation_cancel_open_orders: bool,
@@ -603,7 +603,7 @@ impl BacktestVenueConfig {
     }
 
     #[must_use]
-    pub fn liquidation_trigger_ratio(&self) -> Decimal {
+    pub fn liquidation_trigger_ratio(&self) -> f64 {
         self.liquidation_trigger_ratio
     }
